+++ src/containers/gameController.js
@@ -1,0 +1,4 @@
+// Module that will control the main game loop
+// The game loop should set up a new game by creating Players and Gameboards.
+// For now just populate each Gameboard with predetermined coordinates.
+// You are going to implement a system for allowing players to place their ships later.