+++ src/containers/pipe.js
@@ -1,7 +1,9 @@
+// Articles about loosely coupling object inheritance with factory functions and pipe
 // https://medium.com/dailyjs/building-and-composing-factory-functions-50fe90141374
-export default (...fns) =>
-  (value) => {
-    return fns.reduce((obj, fn) => {
-      return fn(obj);
-    }, {});
-  };
+// https://www.freecodecamp.org/news/pipe-and-compose-in-javascript-5b04004ac937/
+// Observation: if there is no initial value, the first function does not run
+export default (initialFn, ...fns) => {
+  return fns.reduce((obj, fn) => {
+    return Object.assign(obj, fn(obj));
+  }, initialFn());
+};